add : freelancer logic
프리랜서 로직을 추가했고 또한 레파지토리를 작성하지 않은게 생각나서 Entity에 대한 레파지토리도 각각 작성해놨다.

+++ src/main/java/com/bringup/member/resume/domain/entity/CVFree.java
@@ -18,7 +18,7 @@ public class CVFree {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "cv_index")
-    private Long cvIndex;
+    private Integer cvIndex;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_index", nullable = false)
